+++ runtimes/protocol/inlineCompletionWithReferences.ts
@@ -18,13 +18,26 @@ interface DocumentChangeParams {
 }
 
 interface OpenTabParams {
-    openTabFilepaths?: string
+    openTabFilepaths?: string[]
+}
+
+interface FileContext {
+    leftFileContent: string
+    rightFileContent: string
+    filename: string
+    fileUri?: string
+    programmingLanguage: string
+}
+
+interface FileContextParams {
+    fileContextOverride?: FileContext
 }
 
 export type InlineCompletionWithReferencesParams = InlineCompletionParams &
     PartialResultParams &
     DocumentChangeParams &
-    OpenTabParams
+    OpenTabParams &
+    FileContextParams
 
 export const inlineCompletionWithReferencesRequestType = new ProtocolRequestType<
     InlineCompletionWithReferencesParams,